requirements.txt
psycopg2-binary==2.9.9
djangorestframework==3.15.1